--- src/main/java/com/example/namo2/domain/memo/MoimMemoLocationRepositoryCustom.java
@@ -1,9 +1,0 @@
-package com.example.namo2.domain.memo;
-
-import com.example.namo2.domain.moim.ui.dto.MoimMemoLocationDto;
-
-import java.util.List;
-
-public interface MoimMemoLocationRepositoryCustom {
-    List<MoimMemoLocationDto> findMoimMemo(Long moimScheduleId);
-}
