+++ src/main/java/com/example/namo2/domain/memo/application/MoimMemoFacade.java
@@ -1,0 +1,14 @@
+package com.example.namo2.domain.memo.application;
+
+import com.example.namo2.domain.memo.application.impl.MoimMemoService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+@Slf4j
+public class MoimMemoFacade {
+    private final MoimMemoService moimMemoService;
+
+}
