+++ src/main/java/com/example/namo2/domain/memo/dao/repository/MoimMemoLocationRepositoryCustom.java
@@ -1,0 +1,9 @@
+package com.example.namo2.domain.memo.dao.repository;
+
+import com.example.namo2.domain.memo.ui.dto.MoimMemoResponse;
+
+import java.util.List;
+
+public interface MoimMemoLocationRepositoryCustom {
+    List<MoimMemoResponse.MoimMemoLocationDto> findMoimMemo(Long moimScheduleId);
+}
