+++ src/main/java/com/example/namo2/domain/memo/dao/repository/MoimMemoRepository.java
@@ -1,4 +1,4 @@
-package com.example.namo2.domain.memo;
+package com.example.namo2.domain.memo.dao.repository;
 
 import com.example.namo2.domain.memo.domain.MoimMemo;
 import com.example.namo2.domain.moim.domain.MoimSchedule;
