--- src/main/java/com/example/namo2/domain/moim/ui/dto/MoimMemoLocationDto.java
@@ -1,31 +1,0 @@
-package com.example.namo2.domain.moim.ui.dto;
-
-import com.example.namo2.domain.memo.domain.MoimMemoLocationAndUser;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@NoArgsConstructor
-@Getter
-public class MoimMemoLocationDto {
-    private Long moimMemoLocationId;
-    private String name;
-    private Integer money;
-    private List<Long> participants;
-    private List<String> urls;
-
-    public MoimMemoLocationDto(Long moimMemoLocationId, String name, Integer money, List<String> urls) {
-        this.moimMemoLocationId = moimMemoLocationId;
-        this.name = name;
-        this.money = money;
-        this.urls = urls;
-    }
-
-    public void addLocationParticipants(List<MoimMemoLocationAndUser> participants) {
-        this.participants = participants.stream()
-                .map(moimMemoLocationAndUser -> moimMemoLocationAndUser.getUser().getId())
-                .collect(Collectors.toList());
-    }
-}
