--- src/main/java/com/example/namo2/domain/moim/ui/dto/MoimMemoReq.java
@@ -1,12 +1,0 @@
-package com.example.namo2.domain.moim.ui.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@NoArgsConstructor
-@Getter
-public class MoimMemoReq {
-    List<LocationInfo> locationInfos;
-}
